refactor _math
_pydecimal from CPython Lib, but edited
_pymath is only a wrapper for standard library math

+++ pyverm/_math/_pymath.py
@@ -1,3 +1,22 @@
+#=======================================================================
+#=======================================================================
+#  _____                                      _                     _
+# |_   _|                                    | |                   | |
+#   | |    _ __ ___    _ __     ___    _ __  | |_    __ _   _ __   | |_
+#   | |   | '_ ` _ \  | '_ \   / _ \  | '__| | __|  / _` | | '_ \  | __|
+#  _| |_  | | | | | | | |_) | | (_) | | |    | |_  | (_| | | | | | | |_
+# |_____| |_| |_| |_| | .__/   \___/  |_|     \__|  \__,_| |_| |_|  \__|
+#                     | |
+#                     |_|
+#=======================================================================
+#
+# This file is only a facade for the built in math module, but it returns
+# always an Decimal types
+#
+#=======================================================================
+#=======================================================================
+
+
 ########################################################################
 #                                                                      #
 # Copyright (C) 2018,  Marius Hürzeler                                 #
@@ -22,7 +41,7 @@
 decimal math stores the functions to make precise math with decimals
 """
 import math as standard_math
-
+from ._pydecimal import Decimal
 
 
 __all__ = [
@@ -35,91 +54,91 @@ __all__ = [
 ]
 
 
-pi = standard_math.pi
+pi = Decimal(standard_math.pi)
 
 def acos(x):
     """Return the arc cosine (measured in radians) of x."""
 
-    return standard_math.acos(x)
+    return Decimal(standard_math.acos(x))
 
 
 def asin(x):
     """Return the arc sine (measured in radians) of x."""
 
-    return standard_math.asin(x)
+    return Decimal(standard_math.asin(x))
 
 
 def atan(x):
     """Return the arc tangent (measured in radians) of x."""
-    return standard_math.atan(x)
+    return Decimal(standard_math.atan(x))
 
 
 def atan2(y, x):
     """Return the arc tangent (measured in radians) of y/x.
     Unlike atan(y/x), the signs of both x and y are considered."""
-    return standard_math.atan2(y, x)
+    return Decimal(standard_math.atan2(y, x))
 
 
 def cos(x):
     """Return the cosine of x as measured in radians."""
-    return standard_math.cos(x)
+    return Decimal(standard_math.cos(x))
 
 
 def hypot(x, y):
     """Return the Euclidean distance, sqrt(x*x + y*y)."""
-    return standard_math.hypot(x, y)
+    return Decimal(standard_math.hypot(x, y))
 
 
 def sin(x):
     """Return the sine of x as measured in radians."""
-    return standard_math.sin(x)
+    return Decimal(standard_math.sin(x))
 
 
 def tan(x):
     """Return the tangent of x (measured in radians)."""
-    return standard_math.tan(x)
+    return Decimal(standard_math.tan(x))
 
 def exp(x):
     """Return e raised to the power of x."""
-    return standard_math.exp(x)
+    return Decimal(standard_math.exp(x))
 
 
 def expm1(x):
     """Return e raised to the power of x, minus one."""
     # Decimal handles this perfecly so no need for complexity
-    return standard_math.expm1(x)
+    return Decimal(standard_math.expm1(x))
 
 
 def log(x, base=None):
     """log(x[, base]) -> the logarithm of x to the given base.
     If the base not specified, returns the natural logarithm (base e) of x."""
-    return standard_math.log(x, base=None)
+    return Decimal(standard_math.log(x, base=None))
 
 
 def log1p(x):
     """Return the natural logarithm of 1+x (base e)."""
     # Decimal handles this perfecly so no need for complexity.
-    return standard_math.log1p(x)
+    return Decimal(standard_math.log1p(x))
 
 
 def log2(x):
     """log2(x) -> the base 2 logarithm of x."""
-    return standard_math.log2(x)
+    return Decimal(standard_math.log2(x))
 
 
 def log10(x):
     """log10(x) -> the base 10 logarithm of x."""
-    return standard_math.log10(x)
+    return Decimal(standard_math.log10(x))
 
 
 def pow(x, y):
     """Return x raised to the power y."""
-    return standard_math.pow(x, y)
+    return Decimal(standard_math.pow(x, y))
 
 
 def sqrt(x):
     """Return the square root of x."""
-    return standard_math.sqrt(x)
+    return Decimal(standard_math.sqrt(x))
 
 
 # Number-theoretic and representation functions
@@ -127,47 +146,47 @@ def sqrt(x):
 
 def ceil(x):
     """Return the smallest integral value >= x."""
-    return standard_math.ceil(x)
+    return Decimal(standard_math.ceil(x))
 
 
 def copysign(x, y):
     """Return a Decimal with the magnitude (absolute value) of x but the sign of
        y. On platforms that support signed zeros, copysign(1.0, -0.0)
        returns -1.0."""
-    return standard_math.copysign(x, y)
+    return Decimal(standard_math.copysign(x, y))
 
 
 def fabs(x):
     """Return the absolute value of x."""
-    return standard_math.fabs(x)
+    return Decimal(standard_math.fabs(x))
 
 def factorial(x):
     """Return x factorial. Raises ValueError if x is not integral or is
        negative."""
-    return standard_math.factorial(x)
+    return Decimal(standard_math.factorial(x))
 
 
 def floor(x):
     """Return the largest integral value <= x."""
-    return standard_math.floor(x)
+    return Decimal(standard_math.floor(x))
 
 
 def fmod(x, y):
     """Returns the remainder of x and y, using the remainder_near() function
        in Decimal, which comes close to the function in the math library"""
-    return standard_math.fmod(x, y)
+    return Decimal(standard_math.fmod(x, y))
 
 
 def frexp(x):
     """Return the mantissa and exponent of x as the pair (m, e). m is a Decimal
        and e is an integer such that x == m * 2**e exactly. If x is zero,
        returns (0.0, 0), otherwise 0.5 <= abs(m) < 1."""
-    return standard_math.frexp(x)
+    return Decimal(standard_math.frexp(x))
 
 
 def fsum(iterable):
     """Return an accurate floating point sum of values in the iterable."""
-    return standard_math.fsum(iterable)
+    return Decimal(standard_math.fsum(iterable))
 
 
 def gcd(a, b):
@@ -175,63 +194,63 @@ def gcd(a, b):
         If either a or b is nonzero, then the value of gcd(a, b) is
         the largest positive integer that divides both a and b.
         gcd(0, 0) returns 0. """
-    return standard_math.gcd(a, b)
+    return Decimal(standard_math.gcd(a, b))
 
 
 def ldexp(x, i):
     """Return x * (2**i). This is essentially the inverse of function
        frexp()."""
-    return standard_math.ldexp(x, i)
+    return Decimal(standard_math.ldexp(x, i))
 
 
 def modf(x):
     """Return the fractional and integer parts of x. Both results carry
        the sign of x."""
-    return standard_math.modf(x)
+    return Decimal(standard_math.modf(x))
 
 
 def remainder(x, y):
     """Returns the remainder of x and y, using the remainder_near() function
        in Decimal, which comes close to the function in the math library"""
-    return standard_math.remainder(x, y)
+    return Decimal(standard_math.remainder(x, y))
 
 
 def sign(x):
     """Return -1 for negative numbers and 1 for positive numbers."""
-    return standard_math.sign(x)
+    return Decimal(standard_math.sign(x))
 
 
 def signt(x):
     """Return -1 for negative numbers and 1 for positive numbers and 0 for
        zeroes and NaNs."""
-    return standard_math.signt(x)
+    return Decimal(standard_math.signt(x))
 
 
 def acosh(x):
     """Return the inverse hyperbolic cosine of x."""
-    return standard_math.acosh(x)
+    return Decimal(standard_math.acosh(x))
 
 
 def asinh(x):
     """Return the inverse hyperbolic sine of x."""
-    return standard_math.asinh(x)
+    return Decimal(standard_math.asinh(x))
 
 
 def atanh(x):
     """Return the inverse hyperbolic tangent of x."""
-    return standard_math.atanh(x)
+    return Decimal(standard_math.atanh(x))
 
 
 def cosh(x):
     """Return the hyperbolic cosine of x."""
-    return standard_math.cosh(x)
+    return Decimal(standard_math.cosh(x))
 
 
 def sinh(x):
     """Return the hyperbolic sine of x."""
-    return standard_math.sinh(x)
+    return Decimal(standard_math.sinh(x))
 
 
 def tanh(x):
     """Return the hyperbolic tangent of x."""
-    return standard_math.tanh(x)
+    return Decimal(standard_math.tanh(x))
